pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "climate-data-lake-iceberg"
version = "0.1.0"
description = "Real-time Climate Data Analytics Platform with ELT on GCP using Apache Iceberg"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Climate Data Team", email = "example@example.com"},
]
dependencies = [
    "google-cloud-storage>=2.8.0",
    "google-cloud-bigquery>=3.11.0",
    "google-cloud-pubsub>=2.17.0",
    "functions-framework>=3.3.0",
    "pyiceberg>=0.16.1",
    "pandas>=2.0.1",
    "pyarrow>=12.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pytest>=7.3.1",
    "pytest-mock>=3.10.0",
    "argparse>=1.4.0",
    "apache-airflow>=2.6.3",
    "pyyaml>=6.0",
    "gcsfs>=2023.9.0"
]

[project.optional-dependencies]
dev = [
    "black>=23.3.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
    "mypy>=1.3.0",
    "pytest-cov>=4.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = "test_*.py" 